move menu toggling to AuthHeader. User can click avatar or chevron to open menu.
