+++ src/components/navbar/AuthHeader.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { useEffect, useState , useRef} from "react";
 import { getUser, serverSignIn } from "@/app/(auth)/authService";
 import { clientSignIn } from "@/store/features/auth/authSlice";
 import { getUserState } from "@/store/features/user/userSlice";
@@ -14,6 +15,9 @@ export default function AuthHeader() {
   const dispatch = useAppDispatch();
   const { isAuthenticated } = useAppSelector((state) => state.auth);
   const { avatar } = useAppSelector((state) => state.user);
+  const [ isMenuOpen, setIsMenuOpen ] = useState(false)
+  const menuRef = useRef<HTMLDivElement>(null);
+  
   async function handleClick() {
     await serverSignIn();
     const user = await getUser();
@@ -23,14 +27,32 @@ export default function AuthHeader() {
     }
   }
 
+  const toggleMenu = () => {
+    setIsMenuOpen(!isMenuOpen)
+  };
+  const closeMenu = () => {
+    setIsMenuOpen(false);
+  };
+  const handleClickOutside = (event: MouseEvent) => {
+    if (menuRef.current && !menuRef.current.contains(event.target as Node)) {
+      closeMenu();
+    }
+  };
+  useEffect(() => {
+    document.addEventListener("click", handleClickOutside);
+    return () => {
+      document.removeEventListener("click", handleClickOutside);
+    };
+  }, [isMenuOpen]);
+  
   return (
     <>
       {isAuthenticated ? (
         <>
           <Bell notificationCount={notificationCount} />
-          <div className="flex flex-row items-center px-2 ml-6">
-            <Avatar image={avatar} height={34} width={34} />
-            <DropDown openState={false} />
+          <div ref={menuRef} onClick={toggleMenu}className="flex items-center px-2" >
+            <Avatar image={avatar} height={34} width={34}/>
+            <DropDown openState={isMenuOpen}/>
           </div>
         </>
       ) : (
