+++ src/components/navbar/DropDown.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useState, useRef, useEffect } from "react";
+
 import { ChevronDownIcon } from "@heroicons/react/24/outline";
 import DropDownLink from "./DropDownLink";
 import Button from "@/components/Button";
@@ -13,33 +13,13 @@ export default function DropDown({
   openState?: boolean;
 }) { 
   const dispatch = useAppDispatch();
-  const [isOpen, setIsOpen] = useState(openState);
-  const allVoyages = useAppSelector((state) => state.user.voyageTeamMembers) ?? []
-  const activeVoyage = allVoyages.find((item) => item.voyageTeam.voyage.status.name==="Active")
+  const allVoyages = useAppSelector((state) => state.user.voyageTeamMembers);
+  const activeVoyage = allVoyages?.find((item) => item.voyageTeam.voyage.status.name==="Active")
   const currentVoyage = activeVoyage?.voyageTeam.name ?? "Please join a voyage to see your status information.";
-  const menuRef = useRef<HTMLUListElement>(null);
   const closed = "hidden";
   const open =
     "absolute z-[1] w-44 p-4 [&>*]:mt-2 mt-6 shadow bg-base-100 right-0 border border-neutral rounded-2xl";
-
-  const toggleMenu = () => {
-    setIsOpen(!isOpen);
-  };
-  const closeMenu = () => {
-    setIsOpen(false);
-  };
-  const handleClickOutside = (event: MouseEvent) => {
-    if (menuRef.current && !menuRef.current.contains(event.target as Node)) {
-      closeMenu();
-    }
-  };
-  useEffect(() => {
-    document.addEventListener("click", handleClickOutside);
-    return () => {
-      document.removeEventListener("click", handleClickOutside);
-    };
-  }, [isOpen]);
-  
+    
   async function handleClick() {
     await serverSignOut();
     dispatch(clientSignOut());
@@ -50,14 +30,12 @@ export default function DropDown({
       <label
         tabIndex={0}
         className="flex btn m-0 p-0 bg-transparent border-none hover:border-none hover:bg-transparent text-base-300"
-        onClick={toggleMenu}
       >
         <ChevronDownIcon className="w-4 text-base-300 cursor-pointer" />
       </label>
       <ul 
         tabIndex={0} 
-        className={isOpen ? open : closed} 
-        ref={menuRef}
+        className={openState ? open : closed} 
       >
         <li className="bg-secondary-content text-xs p-2 [&>*]:m-1 rounded-lg ">
           <p className="text-neutral text-xs">
